perf(brotli): hierarchical centroid reduction + f32 cost model in the splitter
The flat O(k^2) final combine in cluster_blocks and cluster_histograms
is bandwidth-bound at centroid scale: fits q11's Option-R clustering
evaluated 116M histogram pairs (~19-20s, 81% of the post-task-37
encode); csv2m 24M (~3s). Survivors of the 64-wide batch loop now
re-window in force-merged (<=16 each) batches until <=2048 centroids
remain, then the flat combine runs on the reduced set. The final remap
(every row vs every final cluster) is untouched. fits R 19.4s -> 0.6s,
csv2m 2.96s -> 0.05s.

The splitter's cost pipeline also moves f64 -> f32 with a deterministic
degree-10 Chebyshev log2 (no 512KB table, no libm in the hot loop) —
timing-neutral on its own but removes the cross-platform libm
dependency.

Corpus: q5/q9 byte-identical (small k never enters the reduction);
q11 within +0.1% worst, csv2m -0.868% (better clustering). All
round-trip via the reference decoder.

fits q11 T 3.4 -> 0.59 (faster than the reference at equal load);
csv2m q11 3.26 -> 1.81; words q11 -> 1.18. Board v35.

+++ omnizip-brotli/src/encoder/distance_config.rs
@@ -170,7 +170,8 @@ impl DistanceConfig {
                 if !ok {
                     break;
                 }
-                let cost = crate::encoder::block_splitter::population_cost(&hist) + extra_bits;
+                let cost =
+                    f64::from(crate::encoder::block_splitter::population_cost(&hist)) + extra_bits;
                 if cost > best_cost {
                     break;
                 }
